Added option to bypass signature validation.
